Add vrrp script check for docker service
Prometheus_server_vip relates to the docker swarm cluster,
and as it runs within it, from the keepalive side, we should check
if the docker daemon is working. If on mon node docker service is
down or unresponsive, then the VIP should be moved to other mon node.

Change-Id: Ia5c1e393f38c9d86fa576ead5488fd813b8e20dd
Related-Bug: PROD-18027
Partial-Fix: PROD-17680