Change parameter names for ceph radosgw
There are use cases when backend for rgw lives on separate node, in this case
we have have ability to override it on controller node, as haproxy leaves there.
It was done by https://gerrit.mcp.mirantis.net/#/c/17919/.
However it leads to situation of cyclic assignment of variables, with is not
compatible with 'new' reclass.

To avoid cyclic assignment this patch drops defaults for the followin variables
as they still defined in cookiecutter and static models.
Change-Id: Ib382f2dc6f77a6d02ae7231610c0385be526fe9c