Pin reclass to 1.4.1 on test-salt-model-node jobs
In original reclass it is allowed to override (clear lists/dicts)
by using the following constructions:

a.yml:
  list:
    - a
    - b

b.yml
classes:
  - a
parameter:
  list:
    a~

c.yml
classes:
 - b
parameter:
  list:
    - c

In latest version v1.5.2 and v1.5.3 this is not allowed by default and new syntax
was introduced. The same may be archived by using the following constructions:

a.yml:
  list:
    - a
    - b

c.yml
classes:
  - a
parameter:
  ~list: [d]

We have to pin reclass on model rendering jobs (test-salt-model-node) to latest
version that is present in our repos (as actually this is what our customers are using),
and when new version of reclass is packaged we can smoothly switch to new reclass version.

Change-Id: I68da60764c9b32dbfc809c9d5469082329967dfe
Related-Prod: PROD-20425